IJent: report whole process output during bootstrapping ijent
It's very likely that something important is written to stderr if IJent can't start. Unfortunately, the previous approach used to stop reading stderr as soon as an error appears. The stderr of a remote process comes to the IDE with a delay, hence, important messages used to not be logged.

GitOrigin-RevId: e1af482c752388bf006c4ec87d7e2a8e46431ee2