improve blocked groups handling in new update mode
1. Speed-up action collection in `sendActionRegistry`
By avoiding `getChildren` and extra action instantiation.

2. Fix ActionManagerStateActionItemDescriptor.groupIds in `replaceAction`
Replacement action now keeps the original groups.

3. Honor blocked groups on subsequent registration model arrivals
Blocked actions are not transferred to frontend. Backend actions are not registered for them. Do not register them later too on registration model arrival - keep the original frontend action.

4. Use flows for cancellability

Also,
Fixes `CopyPasteSpecificTest`. Host action `HideToolWindowAction` is blocked via its super-parent `WindowMenu`. That was previously ignored in the new mode two times - in `sendActionRegistry` and in the subsequent re-registration during `asyncActionsModel.rearrange` response processing. The easiest way to fix the test is to add `ActionRemoteBehavior.FrontendOnly` spec to the action, but we need a general solution too.

GitOrigin-RevId: 088d437ef47694b5abc2a87d88d376e31cd83ec1