Remove Performance Testing plugin from AS
Apparently some parts of this plugin are still built from IJ Ultimate
sources only. In particular, we do not have access to the build
scripts for the bundled async-profiler and YourKit profiler, making
this plugin rather incomplete in Android Studio.

If users need this plugin (e.g. for reporting performance bugs) then
they should download it from the plugin marketplace instead (as was
the case before IJ 233).

Bug: 314149266
Change-Id: I2cbb7af24365498e04af2d5c4eb63d3e329a376b