IDEA-330628 Move Compact Mode values from code to theme
Use the Darcula theme as all other themes inherit from it
one way or another. Of course, the Darcula theme isn't supposed
to be used for the new UI, but as these new keys end with
the .compact suffix, they're not used in the old UI,
and don't affect old UI themes directly.

Some keys had to be renamed for their type to be detected
properly, so "insets" should be always at the end.

GitOrigin-RevId: 2cf388786d1030499a72a778451a82ad712af28e