(IDEA-342915) Handle no space left on device exception in LocalEmbeddingIndexFileManager
Since there is no dedicated exception for this, I manually check the message in a caught IOException when saving an index to disk. Another approach would be to check the available space in partition with File::getFreeSpace or File::getUsableSpace, but it looks less reliable than handling the error only when it really happens.

GitOrigin-RevId: 89d53397ddd131877a6b35d041c95636339d7b9e