IDEA-342170 Implement Tree.additionalRowsCount
The implementation is similar to JBTable's, but with some
differences.

The row height is taken from the tree itself, if it's set to a non-zero
value, but with a fallback to the theme default (which is normally
non-zero), as unlike tables we don't really care about computed row
heights, as those can vary greatly in trees, and for the purpose
of creating some extra height we don't want any of that variance,
as it can lead to unpredictable results.

The default value is loaded from the Registry only when actually
computing the height. This is done to avoid early registry access
when the tree is created.

The minimum sensible height is, similarly to tables, set to
the viewport height minus one row, so that there's at least something
visible when the tree is scrolled all the way to the bottom.
However, when the viewport is crazily tiny, even smaller than one row,
we set the extra height to zero. Tables use the full extra length in this
case for some reason, but for trees it definitely isn't going to look good,
though it's really a very edge case that doesn't matter much. But still,
if the viewport is so tiny, we definitely want to see at least something in it.

GitOrigin-RevId: 535fa0b303efdd3dc71657c3c76c85170f40cdd2