EA-474196 Check for EDT in SelectInContextImpl.getElementFromStructureView
If the current thread is not EDT, then this code didn't work anyway
because disposing the structure view component would rightly
throw an exception.

If it's really needed to create a select-in context for a BGT action update,
then the corresponding structure view is better reimplemented as
TreeBasedStructureViewBuilder, just like the vast majority of structure
views. Then a BGT update can access the model directly, without
creating a GUI component.

GitOrigin-RevId: 076a579777abaddf29f0bcd66deb1267234c0715