[git-stash] fix painting branch labels for stashes in the new ui
Instead of painting the label on top of the renderer and calculating the location manually, convert GitStashPainter to a component and place this component in the renderer. This will allow to simplify the code and would fix the problem of needing to paint the label background to cover the main renderer underneath.

GitOrigin-RevId: c32af4b82fc43b11065daf51c9b3c85cc54d40d9