[json] IDEA-339628 Introduced an implementation of the non-root JsonSchemaObject factory
It turned out, that some of the schema nodes are accessed so frequently, that it's worth caching them no matter how lightweight they are. The suggested idea is to have top level schema objects stored in the corresponding JsonSchemaObjectStorage service, and have all the nested nodes inside it. The point of having them all stored in a cache in this factory is that these objects are very often referenced from various schema nodes via '$ref' -> it is convenient to cache and retrieve them by their absolute JsonPointer which is exactly what this factory does.

Some of the created nodes can be additionally stored in lazy properties of JsonSchemaObject instances that appear to be their direct parents. This is done to achieve optimized memory allocation for cases when some additional processing is done with the JsonSchemaObject instances before they are returned to the caller (e.g. they are put into a collection -> we store the collection in a lazy value)

GitOrigin-RevId: 12976e1cbde28228754c2055c01932ef3e1d895b