[FL-23715] improve document changed at slicedCompletion when generating elements
`slicedCompletion` is implemented as a couple of write-actions to instantiate completion session, followed by a series of read-actions when each contributor is called separately, and later collected items are sorted in CompletionLookupArrangers.
It can easily happen, that document is modified somewhere in between these read-actions. In this case the pseudo-inserted psiElement might get invalidated, this CompletionParameters - corrupted.
Before there was a mechanism only at sorting-phase to detect such a corruption and abort current completion;
Now at phase of calling the contributors we also detect such a situation and try to recover - recreate CompletionParameters;
At sorting phase we also detect the specific exception; in case of corruption we send a token to restart the current completion.

GitOrigin-RevId: a44a0da51a1c5cf347b1ae4e6f1bf6e69c60b65f