[workspace model] IDEA-330986 Invalidate global cache at `Restore Default Settings` action call
Setting for SDK and global libraries stored in the configuration folder, so after the `Restore Default Settings` action, this folder cleaned up, so we need to invalidate our global caches. This is the simplest solution to call `invalidateCaches` directly from the action. If there will be a more general solution, e.g., event after the resetting settings, we definitely need to migrate on it, but I didn't find such for now.

GitOrigin-RevId: 579f21e2ee6089c31ebc8aab29e4552d18c750fd