[debugger] Fix exceptions during evaluation of expressions with captured "this" in static lambdas with unnamed parameter
Fix IDEA-340860, the previous commit was an important preparation.

The previous approach was to extract values by the StackFrame#getArgumentValues to get "this" value. But in static lambdas with unnamed parameters, it was leading to exceptions (see JDK-8323747). With JDI workaround (from the previous patch), it became possible to look at "this" parameters directly.

GitOrigin-RevId: 160e748806247d611bbe65c2be8cb4f89b19e6c8