[build scripts] generate a separate bundle for embedded JetBrains Client in macOS distribution (RDCT-921)
If embedded JetBrains Client uses the same executable and same application bundle as the main IDE, macOS activates the client window when you click on the IDE icon if the client is running. To avoid this, now build scripts generate a separate application bundle under Helpers directory, which delegates execution to the main executable.

GitOrigin-RevId: 3483e04246902fb14f254a5472a28262e4c5335e