IJPL-274 no progress reporting from inside `runBlockingCancellable` under indicator
We want to start reporting with a fresh step, and make caller choose how to report (sequential/raw).
Several strategies were considered.
1. Push state, so that the fresh step will report to indicator as if the indicator was fresh, pop state back.
This makes sense, but we'll wait for a use case.
After all, push/pop semantics is one of things which were wrong with `ProgressIndicator` in the first place.
2. Don't report to indicator. The chosen strategy because it's simplest and yet correct.
3. Use amount of space from the current indicator fraction up to the 1.0. Most inconsistent.

GitOrigin-RevId: 070d0b0e11574dd5332fb4c67dec961a3f89efe8