RUBY-32347 check FileBasedIndex class before hard-casting to FileBasedIndexImpl
With the introduction of `EmptyFileBasedIndex`, it is not guaranteed that `FileBasedIndex#getInstance` will return a `FileBasedIndexImpl` while inside of tests (as it can get mocked to return `EmptyFileBasedIndex`). This commit updates a couple of the locations that were being hit by our Ruby/RBS tests to consider this scenario.

(cherry picked from commit 570bac6c92b7753e8560910ecdfc301b99ab81af)

IJ-CR-124206

GitOrigin-RevId: 57b117599750a623d01f94f526a14bafd65ef720