[classpath] IJPL-401 Turn off `classpath.index` by default
Before the caching was enabled in `UrlClassLoader` by default, and it led to problems in different areas (IJPL-401, IDEA-322782, IDEA-305759). So, just simple usage of `UrlClassLoader` without even being aware of this functionality can easily break something, e.g., Gradle project compilation. So, it was decided to turn it off by default and use it explicitly where it's needed.

GitOrigin-RevId: 3212f4b3767a1e05fcf240ce852de7a520ee0997