IDEA-343394 IJent+WSL: propagate `ProcessBuilder.directory` to IJent
There was a false assumption that every code that executes something on WSL sets the working directory via `WSLCommandLineOptions.setRemoteWorkingDirectory`. However, the issue mentioned in the header demonstrates that the git support also sets the working directory via the most usual way, although that way has disadvantages in WSL particularly.

GitOrigin-RevId: 764a6f7e1809b223062779b2f7c3ba71ac835814