IDEA-339254 Don't use the base symbol icon instead of the actual one
An old bug introduced in 2011 when refactoring JSP stuff: getClassIcon
may accept a non-null argument to force a specific icon (used only by JSP).
However, the same argument is also assigned a non-null value of the base
symbol icon which obviously shouldn't replace the actual one when the actual
one is evaluated. But it exactly what happened, because the value was assigned
to the same argument that is used to force an icon.

Fix by using a proper local variable instead of assigning to the argument.
Now the icon will only be forced if it's explicitly passed in by the caller,
and in other cases it'll be properly evaluated when the time comes,
replacing the base one as it should.

GitOrigin-RevId: b37a6d464a34262b29474e489c54e63906dc3d7a