IDEA-343180 Enable execution process in WSL with IJent
The registry flag wsl.use.remote.agent.for.launch.processes becomes true

All processes that are started in WSL through WSLDistribution.patchCommandLine will be running through IJent.

IJent is a new daemon running on a remote machine that communicates with an IDE via RPC. IJent takes control on running processes on WSL, including a shell in a terminal.

GitOrigin-RevId: beedd39ef1743f5c7875c59e6c18362dffad01d0