(IJPL-515) Search Everywhere ML experiment groups shuffling strategy
SearchEverywhereMlExperiment class now has a new static property VERSION used together with 'device_id' when we determine an experiment group. Thus, within one experiment version, the experiment group does not change, but it does change in a uniform way when we increment the version. So in new experiment groups, every old experiment group should present equally or almost equally. I also get rid of a separate experiments manager for semantic search.

To preserve access to an experiment group from the FUS collector (which we could previously achieve with the column 'bucket'), I start reporting it from the SearchEverywhereUsageTriggerCollector

GitOrigin-RevId: 29f98f20a0f91c1ca772453e706f5d3eef460ae0