IDEA-339241 Process action shortcuts in popups
Simply create a list of actions and register their shortcut sets to correspond
to the list items that have shortcuts.

Actions are used here instead of reacting to shortcuts directly because
it allows to use the standard shortcut processing procedure that
takes into account all numerous gotchas and even allows to process
multi-stroke shortcuts properly.

GitOrigin-RevId: 9813042e1a6c33f5a21a05831ec5d5976c51fadb