IDEA-324815 Move CachingTreePath from impl to api
It's a very useful and harmless optimization that could
be used in more places than it is already used in.

It's especially useful for the new bulk expand/collapse API,
as due to its bulk nature it may have to operate with a lot
of paths, and they're used as hash map keys, which means
there will be a lot of equals() calls, and those compare
path counts among all things.

GitOrigin-RevId: 3aba6745baba92e48bcca91c82fcdaa1fb426739