IDEA-324815 Fix visitThread()==BGT handling by EDT models
Turns out there was a bug in the contract between EdtBgtTreeVisitor
and its callers. The contract assumed that the visit() method will be
called on the EDT or a BGT depending on what visitThread() returns,
and at the same time it was implied that visit() should invoke
preVisitEDT/postVisitEDT directly if it's called on the EDT.

However, pure EDT models, such as the DefaultTreeModel,
and in fact pretty much anything else than AsyncTreeModel,
are blissfully unaware of the fact that the visitThread() method
even exists, and call only the visit() method directly, on the EDT,
of course.

Due to this inconsistency, the pre/post visit methods weren't even
called in these cases. As a result, TreeUtil.promiseExpand stopped
working on these models if its visit thread was set to BGT.

Fix by not checking visitThread() in visit(), but checking the actual
calling thread instead. If it's a BGT, assume that the caller knows
what it's doing, otherwise do all the stuff directly. Implement this
logic directly in EdtBgtTreeVisitor, make it an abstract class, make that
method final and let subclasses implement the new doVisit() method
instead. This enforces the implementation logic on all subclasses
(luckily there's only one at the moment).

GitOrigin-RevId: 156adf6a9ef8b6440c9aed1c75800dfd2b715aab