IJent: send to EA last stderr logs if process exits suddenly
There already were few reports that the IJent process exited with the code 126, but it's turned out to be absolutely unclear why. There was no possibility to find the reporters and ask them to collect more debugging info.

This commit adds an ability to send several latest lines of the stderr of the IJent process if it exits suddenly.

GitOrigin-RevId: 9bf7b8c921020a573403d9c82f998579a61f1f56