[devkit] implement incremental compilation for runtime module repository (IDEA-335089)
The runtime module descriptors file should be updated if and only if something changes in the configuration of modules and libraries. Since up-to-date checks in org.jetbrains.jps.incremental.BuildOperations are based on files, we need to have at least one source file registered by the builder for that. Here .idea/modules.xml is used for that.

GitOrigin-RevId: ca5e56dfca75d63e4056017a4164dd3fbb20b116