[debugger] IDEA-338723 Use thread policy to filter breakpoints for Run-To-Cursor in coroutines
Suspend all (default) policy flaky leads to triggering the same breakpoint on several threads. So it would be needed to check which threads reach needed breakpoint. Alternatively, it is possible to use thread policy for breakpoints.

Need to improve the UI/UX of debugger after this change.

Fix IDEA-344122 KTIJ-28623 IDEA-344124 KTIJ-28623

GitOrigin-RevId: 0830bdc706591cdb71916899fd07211e7333a096