[kotlin] K2 J2k: Port ConvertToIsArrayOfCallFix to a JK conversion
Context from the spreadsheet:
"KTIJ-26655. Looks like ConvertToIsArrayOfCallFix can be just ported to a JK conversion"

NB: The test `CheckForInstanceOfErasedArrayType` fails because the parameters are still treated as non-nullable. Should this step remain as a postprocessing, since it seems like nullability analysis needs to happen first?

Contents of the failed test:
```
class J {
    fun testObject(obj: Any): Boolean {
        return obj is Array<*> && obj.isArrayOf<String>()
    }

    fun testArray(array: Array<Any?>): Boolean {
        return array.isArrayOf<String>()
    }
}
```