IDEA-324815 Fix expand all in JSProjectStructureTest.testExternalLibs
Apparently it's the only test that actually makes use of expand ALL,
including the External Libraries node (which is actually the one
it tests). It doesn't make any sense to always expand this node
for all tests (as it's very often huge), so add a special expandAll()
overload just for this case to allow passing a custom predicate.

GitOrigin-RevId: 21802a6b8fb1b5047e510052ce93a46e38ecb94c