IDEA-343790 Change default Tool Window Layouts group text
Make it more readable in places like Search Everywhere,
but keep it as it is in the main menu.

For some reason, override-text doesn't work for groups unless
there's a class with an overridden update(), so create an empty class
as a workaround.

To avoid confusion with the subgroup that only contains layouts
themselves, but no the other actions (save current, restore, etc.),
remove this inner group from Search Everywhere and rename it
to Tool Window Layout List.

GitOrigin-RevId: 993641ef608cd5d919edc19638d88eaf8da9a4cb