IDEA-343992 Fix highlighting update for inactive windows
The existing code only worked for the case when the window
containing the editor is active. However, there can be multiple
windows per project, and some other window of the same project
may be active.

In this issue specifically we've run into a situation when the active
window is a detached Project View. If editors are opened with
single clicks, the Project View window remains active, and the editors
don't receive proper highlighting.

Fix by improving the existing optimization: try to guess the project
for both the editor and the active window. It's usually easy if the
window implements IdeFrame, and if it doesn't, it's owner typically does.
If the active window and the editor window have the same project,
it means that we still need to update it, even if the editor's window
isn't active.

GitOrigin-RevId: dea66b8658208b5252f8b8d7a29977f1414e1ddd