RDCT-987 Fix "Project-level session is not set" exception on project opening
Project services in RD rely on the fact that if they have been invoked with not local ClientId remote project session (`ClientProjectSession`) already exists for the provided project.

However, we create the session asynchronously on the background thread, so there can be a situation when a project is already initialized, but the remote session hasn't been created yet.

We had quite a lot of these exceptions in `BackendServerToolWindowManager.getToolWindow` and `ControllerNotificationRouter.routeNotification`, because some operations with tool windows and notifications can be executed quite early.
And even if those two are fixed, there could be many more such cases.

We should be able to create a remote session as early as possible, even before any project service can be created.

To achieve this, we need:

1. Use `ProjectServiceContainerInitializedListener` extension.

Unlike `InitProjectActivity` this one is executed before any components or services are created, there-as the second is executed always after.
We use the same extension in JetBrainsClient (frontend part) - `ThinClientProjectExtensionsConfigurator`, so it seems reasonable to use the same tool both on backend and frontend.

2. Remove `launchBackground` calls

To avoid races, we need to create a remote session without scheduling its execution to a background thread.
Using RD protocol requires EDT, so we need to use `withContext(EDT)` (without `launch` to avoid parallel execution)

However, preloading services and staring delivering messages might be time-consuming, so we can launch this part on the background - session is created; everything else can be done later.

GitOrigin-RevId: 8918289b326ae6f5657ed5d8ad1976edb7dfd67d