IJPL-149878 Implement MultiRoutingFileSystem
The new filesystem allows routing specific paths to some backend filesystems.

The design of the new filesystem supposes that it is set as the default nio file system via `java.nio.file.spi.DefaultFileSystemProvider`.

It is going to be used on Windows for routing `\\wsl*` paths to `IjentNioFileSystemProvider` and routing all other paths like `C:\` to the original default file system.

GitOrigin-RevId: 69b5b846f3efe3a03cd373a68ab2fb3fe8a4e573