IDEA-344781 Only announce collapse for topmost nodes in Tree
Now that we have a suspend/resume accessibility announcements API,
let's start using it by suspending useless collapse announcements
when a lot of nodes are collapsed, and only announce collapsing
of the topmost nodes afterward. Recursive collapsing is nothing
more than an implementation detail, after all, and it's not directly
visible when a node is collapsed, so why announce everything?

GitOrigin-RevId: c66056480c2f22e8f8b0efbc9e65b1796f232dc6