[vfs] EA-236640, EA-966425: ensure MMappedFileStorage allocates-zeroes new pages in a crash-tolerant way
File expansion-and-zeroing during mmapping a new page is non-atomic -- it could be interrupted by app crash/kill. This leaves a mmapped storage in an inconsistent state -- either file size != N*pageSize (EA-236640, EA-966425, ...), or the new region is not fully zeroed (these bugs are harder to point exactly, but many errors in VFS could be -- at least partially -- explained by the garbage in a storage).
Tried to make mmapped storage file expansion crash-tolerant.

GitOrigin-RevId: 4385cb7b84d2689dc68f935da68e8aa793b0228c