IDEA-324815 Tree: don't try to expand leaves
JTree.expandPath() checks whether the path is a leaf before expanding it.
Do the same thing in our bulk API. It's not very important in general,
as for leaves it doesn't matter whether it's expanded or not,
but we can save a lot of RAM in expandState this way, as there
can be many leaves, and they're never actually expanded even if
we tell them to.

GitOrigin-RevId: 997be36336a8c5c3fa96cdf27f2f3609ade4d537