[runtime module repository] properly resolve $PROJECT_DIR$ macro in resource paths
Some tests module depends on JAR files located in project sources, so we need to support $PROJECT_DIR$ macro at least to run tests (RDCT-773). It's not correct to suppose that the current directory is the project root, so now the code starts with the base directory for the module repository, and walks up to find the project root.

GitOrigin-RevId: 11d01f5dfb331b77352fb8169a49ce1941206d26