[terminal] IDEA-345156 Fix showing all commands as an error after failed application command
Previously we relied on the `LastExitCode`. This variable is updated when the external application is finished with its exit code. But this value is persisted during the execution of other internal terminal commands (powershell cmdlets and functions, for example). So as a result, we have taken it into account during the calculation of the exit code for the current command while it is outdated. Do not use it at all for now, because `$?` variable is enough now.

GitOrigin-RevId: ea40305f9f0bac92351b33627f2da22433363879