IDEA-344949 Autoscroll to source: use the navigation service if possible
OpenSourceUtil contains a runWithModalProgressBlocking call,
which blocks input events, sometimes blocking the subsequent
click which leads to a double-click being ignored completely.

This doesn't fix the issue fully because there are deeper
runWithModalProgressBlocking calls that are impossible to
get rid at this level. Specifically, in FileEditorManagerImpl.openEditorImpl.
But this fix does help a bit.

GitOrigin-RevId: c466eb5c7dd672edee9e08abbcab5fd2424402a7