IDEA-324815 Use bulk expand in TreeState for non-async models
We can't use bulk expand for async models yet because collecting
paths to expand is too slow, but we can do it already for non-async
models, where all data is already available in the model, and anyway
they were expanded in a single EDT event, causing huge freezes
if the model was unexpectedly large (e.g. the changes tree
in the Commit tool window when there are a hundreds of thousands
of changed/added/deleted files).

Using bulk expand we can fix those freezes or at least turn them
from perma-freezes into micro-freezes in the worst cases.

GitOrigin-RevId: 1d9c9beeb95e479c3bbd9f3ee45295761205d486