IDEA-345148 Implement Editor Gutter Compact Mode
Performance note: as area width suppliers are called VERY
often, instead of using JBUI.CurrentTheme as a supplier directly,
we cache those values in the value group already existing
in EditorGutterComponentImpl for similar purposes.
The only trick is to properly update and re-cache values
on LaF change (Compact Mode toggle), which isn't possible
to do in JBValueGroup because the LaF Manager isn't available
in that module. To get around it, explicitly clear the cache
in EditorGutterComponentImpl.updateUI, which will
lead to the cache being cleared and rebuilt multiple times
if there are many editors, but that will only happen during
LaF change anyway, so it's still better than retrieving values
from JBUI.CurrentTheme over and over again.

GitOrigin-RevId: 8a7c67a072bd7aa1ae2765d71bf5e7aeaf6533d0