IDEA-342509 Improve DefaultTreeUI.updateCachedPreferredSize performance
We have freezes related to delegating to the super calls in this method.
Now that ide.tree.experimental.preferred.width applies to all trees,
this delegation should never ever happen ideally.

Apparently it happens at tricky moments when the parent hierarchy
doesn't provide enough information about sizes, like when it's in
the process of creation or validation. Instead of delegating to super,
which potentially calculates the size of the whole tree, leading to freezes,
we introduce some sensible fallbacks for these cases.

This should fix at least all freezes related to super() calls.
If there are still some more, then we'll have to fix those locally,
as it'll be purely our platform code then.

GitOrigin-RevId: 7f3c730540dbf1d9e45514d3ff647e125e6f7191