IJPL-639 don't request `PluginUtil` instance during `<clinit>`
Typical usage pattern means that `ID.create` is called during `<clinit>`:
```
private static final MY_ID = ID.create(...);
```
This is incorrect because `ID.create` requests a service, and that request may be canceled,
which will forever break the class and result in `NCDFE`.

A service is needed to compute the owner class differently in JDK < 11 and JDK >= 11.
Instead, this logic is moved to `Java11Shim`, which is used in newly added `ID.getCallerPluginId`.

GitOrigin-RevId: 414487f55f9752694019a1fb8e15cf385b44b6eb