IDEA-18765 Cache subpackages for the Packages Project View
When computing children of a node in the Packages view,
the slowest part is getting the list of subpackages of a given
package. Sometimes it happens twice: once to check
if the package is empty, then to actually retrieve the list
of subpackages to create its children.

Fix this part by moving the relevant stuff from PackageUtil
to the new PackageNodeBuilder class which caches packages internally.
Since this is volatile PSI stuff, we can't cache them between updates,
but at least we can cache them during a single node creation,
which already makes it twice faster, though still horribly slow.

GitOrigin-RevId: adadb24259e43e12a167c4c758a9fb4797058533