IDEA-18765 Make modules in the Packages View always non-leaves
The slowest part of a module node update is to figure out if it's a leaf.
This essentially retrieves and validates all of its subpackages,
to figure out if there are any valid ones.

This is only done so that an empty module doesn't have an expand
icon next to it in the tree. Looks like a horrible trade-off, especially
since empty modules are a very rare beast, and displaying an expand
icon next to one isn't really a big deal.

GitOrigin-RevId: 79c0fce83fe309e545224bc1f15f94d9551eeec5