[RDCT] Flush user input immediately, needed for RDCT-5
Before this commit, user input was flushed almost immediately, but postponed with a queue in `print` method. It was `flushDeferredText` some time ago, but was rewritten to support invocation from background thread. Since we know that we are on the dispatch thread in `type` method, we can call `flushDeferredText` here directly.
This change is important for RDCT, and RDCT-5 in particular, to properly track document changes in the same stack they were triggered.

GitOrigin-RevId: c050623cbefd88c024fa99d76d768165b377f083