RDCT-940: remove default `id` property calculation represented by implementation class name
in case of VirtualFileCustomDataSynchronizer different classes are used for consumers and providers, so the results of such calculations will be different, however they must be equal for performing proper data transferring.
in case of ProjectCustomDataSynchronizer it is not a problem yet, because implementations of consumers and providers are united, but it is better to make the same change for consistency.

(cherry picked from commit 6be2d5ece294ae280ef6759d1fc341c8ef6dce69)

IJ-CR-125844

GitOrigin-RevId: d8fb0ce0073135fd7255c6299d5423154703a636