[file-history] implement branch filter in file history without indexes
This commit reworks VcsLogFileHistoryHandler to accept a filter collection and to specify which filters are supported. Handler can also throw a UnsupportedHistoryFiltersException when passed filters are not supported.

Currently, GitFileHistory can not accept starting points and filters separately, so branch heads passed as starting points and hash is ignored. Because of this, filtering by branches is disabled for history for revision.

GitOrigin-RevId: d8232b71419587a62d6d1487a073f172d14acc8f