IDEA-344769 Account for invisible insets for tool window floating bounds
Now that we have a separate property for external decorator bounds,
implement it for both decorators to account for invisible insets.

As there seems to be no API for getting those invisible insets,
we use guesswork: the top inset is usually the window header,
and it's visible; the other insets are usually invisible drag zones.

Insets aren't available until the window is connected to an actual display
(the peer is created), so the new property will work incorrectly before
the window is shown for the first time, but thankfully we already have
a helper class to fix the insets after the window is shown, so it works
as expected, even though it's still a dirty hack and not a proper fix.

GitOrigin-RevId: efd75de13d070a6873bda05898803cc533259194