IDEA-344395 IJent: blocking API for Java uses `runBlockingCancellable` and implicitly requires a progress indicator
Communication with IJent may freeze. No matter how good or bad the code is polished, it's a separate process that communicates through a third-party application (wsl.exe). It may freeze by its nature.

Because the functions used to call `runBlocking`, it could have blocked the calling thread for unpredictable time. Sometimes it was not possible to close the project due to that.

Now functions `fetchLoginShellEnv` and `runProcessBlocking` keep track of the current progress indicator and log an error if there's no progress indicator.

Q: The problem is in the inability to close the project, and there's an argument `project`. Why not hook on it?
A1. De-facto, this argument is always `null`, and it's uneasy to wire the correct project to all invocations of these functions.
A2. It would be a fix of a particular symptom, not the problem. These functions are API, and they may block something else.

Q: There are many error messages after this commit. Why did you break everything?
A: They're just log messages, they don't prevent any calculation or execution. These messages are important for getting to know about places that still lack a progress indicator and may freeze therefore.

GitOrigin-RevId: 8905f11cd0a4e98454bf60d766249c99c7d21d5f