IDEA-346174 Fall back to the actual file extension for Project View sorting
When sorting by type, the default file extension for that type is used
as the sort key. Which is correct in the case when the type defines
a default icon, and allows to group files of the same type together.

However, when several types don't define a default icon (it's an empty
string), it leads to files of different types being mixed up.

Fix by falling back to the actual extension in the case when the default
one is empty. This way we can at least separate unrelated files.

GitOrigin-RevId: f35fb6468dc5844659f93d75b8bd5434b8c40a4c