[statistics] change how information about the product is represented in remove dev cases (RDCT-735)
Now 'product_mode' property is added to statistics data with values 'frontend' for a frontend variant (JetBrains Client) of an IDE and 'backend' for IDE running in 'remote dev host' mode.
A frontend variant of an IDE uses the product code of the corresponding full IDE to report statistics instead of using the generic JetBrains Client product code ('JBC').

GitOrigin-RevId: 6a2d44c98471b4852812bf138894baf73b4b203f