[gitlab] Make private forks checkout through special ref (IDEA-326553)
Fixes IDEA-326553

refs/merge-requests/{iid}/head is used here.
Changes to other places where getSourceRemoteDescriptor is called
should still follow sometime.

Currently, the special ref is checked out as a local branch named fork/refs/merge-requests/{iid}/head.
This is not ideal for a few reasons, but future commits should solve this:

Checking out as a branch makes it possible to confuse the special ref as a remote branch. It's not a remote branch and it cannot be made one. You shouldn't be able to push to it, but pull might be made possible in some special fetch way.

GitOrigin-RevId: c74b15971b2620e707a523206fb4ec9ae1f95d70