[Workspace Model] [IDEA-333373] Move reloading of iml files into the IO dispatcher
On external change of configuration files (`iml`/`xml` files) we have to reload the project model.
Previously, it worked fully on EDT what caused UI freeze. However, there is no requirement to do it on EDT, so now it's moved to the IO dispatcher, removing the part of the freeze.

Unfortunately, the freeze is not removed fully because the update of the configuration store is still under EDT.

Also, the background progress bar was added to indicate that the reloading is in progress

No new tests were added because
`ModuleStoreTest.one batch update session if several modules changed` checks this functionality

GitOrigin-RevId: 155da36116b92c011e00f1edffd562499dbc1f06