[rdct tests] allow using module-based loader to start the client process in tests (RDCT-773)
If system property 'rdct.test.use.modular.loader.for.client' is set to 'true', distributed tests runs JetBrains Client process using the same module-based loader as in production, with per-plugin classloaders, instead of using the same classloader to load all classes. Modules required for tests are added to a special test plugin, which is put to the custom plugin directory.

GitOrigin-RevId: b5e8f85bd15bb6e8964a11241ad42a478dd37065