IDEA-346297 [Gradle|DataNode] fix: instantiate the Gradle module data cache only once
The usage of CachedModuleDataFinder's old implementation has O(N*N) complexity in the worst scenario, because it initialized cache only for taken elements. But cache miss causes a new traversal for all modules. So, if the code collects module data for all modules sequentially, then we have cache miss on each call.

GitOrigin-RevId: 59efdac9d0b3b497d9ce767dcb79d8d9ccf8e1ee