[lvcs] do not match ChangeSets by labels when collecting local history information
If a ChangeSet has some changes, it should be shown in the local history if these changes affect the target file. Labels on the other hand can't be attributed to a specific file and are set on the whole project. So if a ChangeSet has labels as well as changes, labels should be ignored, and only changes should be considered. Otherwise, items without relevant changes are going to appear in the local history, which can be confusing.

If a ChangeSet has only one label and no changes, it can be shown if it was set during the file lifetime.

GitOrigin-RevId: 1791d1d50c308fe27d5ce038a0c18ebe1f304599