[build scripts] ensure that runtime module repository is always prepared
Remote development tests will use the runtime module repository to run JetBrains Client (RDCT-773), and there are plans to use it in other tests as well. So build scripts are modified to ensure that the module repository is built during project compilation. If the build scripts are configured to reuse classes compiled by the IDE, there is no need to build the module repository explicitly because it's generated by the IDE as well.

GitOrigin-RevId: ef856fe9227ecf436b0d97ac9f44e0e8e0d8079e