[terminal] Do not send command started event for generators in Bash
It is interesting that the logic in the `bash-preexec`, that calls our `__jetbrains_intellij_command_started` function is passing the first command from the history as a parameter.
So as a result, when our function is called, we got the previous command instead of the one should be run now.
And because of that, we sent the command started event for generator commands, because the `__jetbrains_intellij_is_generator_command` checks the previous command.
Returned the `BASH_COMMAND` property, that was here before `bash-preexec` is added.

GitOrigin-RevId: e33df3c1cc4225a684f7d3afbb80710293a53172