[terminal] IDEA-344689 Remove Ctrl+C shortcut from copying block on Windows and Linux
It is expected that Ctrl+C will always interrupt the current command, instead of doing something else.
Also, return the focus to the prompt when command is interrupted if any block was selected. It is done by clearing the selection, so when prompt is shown, it will be the preferred focusable component.

GitOrigin-RevId: c364a10e678d8112ccf14e176fae811646ab6a4b